Rename module from sam_cli_entrypoint_adapter to cli_entrypoint to match SAM plugin discovery convention
