+++ cli-entrypoint/logging.yaml
@@ -12,7 +12,7 @@ handlers:
     formatter: standard
     stream: ext://sys.stderr
   file:
-    class: sam_cli_entrypoint_adapter.logging_utils.MkdirRotatingFileHandler
+    class: cli_entrypoint.logging_utils.MkdirRotatingFileHandler
     level: DEBUG
     formatter: standard
     filename: logs/sam-cli-entrypoint.log
