+++ cli-entrypoint/src/cli_entrypoint/__init__.py
@@ -1,0 +1,1 @@
+"""CLI Entrypoint - Terminal-based entrypoint for Solace Agent Mesh."""
