--- cli-entrypoint/src/sam_cli_entrypoint_adapter/__init__.py
@@ -1,1 +1,0 @@
-"""SAM CLI Entrypoint Adapter - Terminal-based entrypoint for Solace Agent Mesh."""
